Update codebase to java 8 style

+++ src/main/java/net/ttddyy/dsproxy/listener/logging/RegisterOutParameterValueConverter.java
@@ -21,7 +21,7 @@ public class RegisterOutParameterValueConverter implements ParameterValueConvert
         sb.append("OUTPUT(");
         // for the second argument, it is either int or SQLType(in JDBC 4.2)
         if (sqlType instanceof Integer) {
-            String sqlTypeName = ParameterValueConverter.SQL_TYPENAME_BY_CODE.get((Integer) sqlType);
+            String sqlTypeName = ParameterValueConverter.SQL_TYPENAME_BY_CODE.get(sqlType);
             if (sqlTypeName != null) {
                 sb.append(sqlTypeName);
                 sb.append("[");
